+++ packages/core/src/types/index.ts
@@ -2,6 +2,7 @@ export * as Collector from './collector';
 export * as Data from './data';
 export * as Destination from './destination';
 export * as Elb from './elb';
+export * as Flow from './flow';
 export * as Handler from './handler';
 export * as Hooks from './hooks';
 export * as Mapping from './mapping';