+++ packages/web/destinations/meta/src/examples/mapping.ts
@@ -13,7 +13,7 @@ export const Purchase: DestinationMeta.Rule = {
           'nested',
           {
             condition: (entity) =>
-              isObject(entity) && entity.type === 'product',
+              isObject(entity) && entity.entity === 'product',
             map: {
               id: 'data.id',
               quantity: { key: 'data.quantity', value: 1 },
@@ -25,7 +25,7 @@ export const Purchase: DestinationMeta.Rule = {
       num_items: {
         fn: (event) =>
           (event as WalkerOS.Event).nested.filter(
-            (item) => item.type === 'product',
+            (item) => item.entity === 'product',
           ).length,
       },
     },
@@ -64,7 +64,7 @@ export const InitiateCheckout: DestinationMeta.Rule = {
           'nested',
           {
             condition: (entity) =>
-              isObject(entity) && entity.type === 'product',
+              isObject(entity) && entity.entity === 'product',
             map: {
               id: 'data.id',
               quantity: { key: 'data.quantity', value: 1 },
@@ -75,7 +75,7 @@ export const InitiateCheckout: DestinationMeta.Rule = {
       num_items: {
         fn: (event) =>
           (event as WalkerOS.Event).nested.filter(
-            (item) => item.type === 'product',
+            (item) => item.entity === 'product',
           ).length,
       },
     },