+++ packages/web/sources/browser/src/__tests__/edgeCases.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-disabled-tests */
 import type { WalkerOS, Collector } from '@walkeros/core';
 import { createCollector } from '@walkeros/collector';
 import { createBrowserSource } from './test-utils';