+++ packages/server/destinations/meta/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/server-destination-meta
 
+## 0.1.0
+
+### Minor Changes
+
+- fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @walkeros/server-core@0.1.0
+  - @walkeros/core@0.1.0
+
 ## 0.0.8
 
 ### Patch Changes