+++ packages/web/destinations/piwikpro/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/web-destination-piwikpro
 
+## 0.1.0
+
+### Minor Changes
+
+- fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @walkeros/web-core@0.1.0
+  - @walkeros/core@0.1.0
+
 ## 0.0.8
 
 ### Patch Changes