+++ packages/web/sources/browser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/web-source-browser
 
+## 0.1.0
+
+### Minor Changes
+
+- fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @walkeros/collector@0.1.0
+  - @walkeros/web-core@0.1.0
+
 ## 0.0.10
 
 ### Patch Changes